tests/wpt/tests/event-timing/orphan-keydown.html

<!DOCTYPE html>
<html>
<meta charset=utf-8 />
<title>Event Timing: orphan keydown.</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/resources/testdriver.js></script>
<script src=/resources/testdriver-actions.js></script>
<script src=/resources/testdriver-vendor.js></script>
<script src=resources/event-timing-test-utils.js></script>

<body>
  <button id='target'>Click me</button>

  <script>
    let observedEntries = [];
    const map = new Map();
    const events = ['keydown'];

    promise_test(async t => {
      assert_implements(window.PerformanceEventTiming, 'Event Timing is not supported.');

      const callback = (entryList) => {observedEntries = observedEntries.concat(entryList.getEntries().filter(filterAndAddToMap(events, map))); };
      const readyToResolve = () => { return observedEntries.length >= 1; };
      const observerPromise = createPerformanceObserverPromise(['event'], callback, readyToResolve);

      await interactAndObserve('orphan-keydown', document.getElementById('target'), observerPromise);

      assert_equals(observedEntries.length, 1, "Keydown without a keyup should be fired properly.");
      assert_greater_than(map.get('keydown'), 0, "Should have a non-trivial interactionId.")
    }, "Event Timing: Orphan keydown should be measured as an interaction.");
  </script>
</body>
</html>
